requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
openpyxl>=3.1.0
reportlab>=4.0.0
numpy>=1.24.0
